Admin/Announcements endpoints (#72)
* List announcements
* Create announcement
* Delete announcement
* Update announcement

Extra:
* All Limit should be `uint` because they are all `0 < Limit <= 100`
* All Offset should be `uint64` because they are all `0 < Offset` and
  they can be large numbers.

Resolves #60

+++ services/antennas/notes.go
@@ -13,7 +13,7 @@ const (
 // NotesRequest represents a request to fetch notes for a given Antenna.
 type NotesRequest struct {
 	AntennaID string `json:"antennaId"`
-	Limit     uint64 `json:"limit"`
+	Limit     uint   `json:"limit"`
 	SinceID   string `json:"sinceId"`
 	UntilID   string `json:"untilId"`
 }
